refactor llm and db into chat folder, refactor their tests into chat folder, and update docs rst
