+++ docs/pykoi.chat.db.rst
@@ -1,0 +1,53 @@
+pykoi.chat.db package
+=====================
+
+Submodules
+----------
+
+pykoi.chat.db.abs\_database module
+----------------------------------
+
+.. automodule:: pykoi.chat.db.abs_database
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.db.comparator\_database module
+-----------------------------------------
+
+.. automodule:: pykoi.chat.db.comparator_database
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.db.constants module
+------------------------------
+
+.. automodule:: pykoi.chat.db.constants
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.db.qa\_database module
+---------------------------------
+
+.. automodule:: pykoi.chat.db.qa_database
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.db.ranking\_database module
+--------------------------------------
+
+.. automodule:: pykoi.chat.db.ranking_database
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pykoi.chat.db
+   :members:
+   :undoc-members:
+   :show-inheritance:
