+++ docs/pykoi.chat.llm.rst
@@ -1,0 +1,69 @@
+pykoi.chat.llm package
+======================
+
+Submodules
+----------
+
+pykoi.chat.llm.abs\_llm module
+------------------------------
+
+.. automodule:: pykoi.chat.llm.abs_llm
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.llm.constants module
+-------------------------------
+
+.. automodule:: pykoi.chat.llm.constants
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.llm.huggingface module
+---------------------------------
+
+.. automodule:: pykoi.chat.llm.huggingface
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.llm.instruct\_pipeline module
+----------------------------------------
+
+.. automodule:: pykoi.chat.llm.instruct_pipeline
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.llm.model\_factory module
+------------------------------------
+
+.. automodule:: pykoi.chat.llm.model_factory
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.llm.openai module
+----------------------------
+
+.. automodule:: pykoi.chat.llm.openai
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.chat.llm.peft\_huggingface module
+---------------------------------------
+
+.. automodule:: pykoi.chat.llm.peft_huggingface
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pykoi.chat.llm
+   :members:
+   :undoc-members:
+   :show-inheritance:
