+++ docs/pykoi.chat.rst
@@ -1,0 +1,19 @@
+pykoi.chat package
+==================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pykoi.chat.db
+   pykoi.chat.llm
+
+Module contents
+---------------
+
+.. automodule:: pykoi.chat
+   :members:
+   :undoc-members:
+   :show-inheritance:
