--- docs/pykoi.db.rst
@@ -1,29 +1,0 @@
-pykoi.db package
-================
-
-Submodules
-----------
-
-pykoi.db.qa\_database module
-----------------------------
-
-.. automodule:: pykoi.db.qa_database
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pykoi.db.ranking\_database module
----------------------------------
-
-.. automodule:: pykoi.db.ranking_database
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pykoi.db
-   :members:
-   :undoc-members:
-   :show-inheritance:
