+++ docs/pykoi.interactives.rst
@@ -1,0 +1,21 @@
+pykoi.interactives package
+==========================
+
+Submodules
+----------
+
+pykoi.interactives.barchart module
+----------------------------------
+
+.. automodule:: pykoi.interactives.barchart
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pykoi.interactives
+   :members:
+   :undoc-members:
+   :show-inheritance:
