--- docs/pykoi.llm.rst
@@ -1,61 +1,0 @@
-pykoi.llm package
-=================
-
-Submodules
-----------
-
-pykoi.llm.abs\_llm module
--------------------------
-
-.. automodule:: pykoi.llm.abs_llm
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pykoi.llm.constants module
---------------------------
-
-.. automodule:: pykoi.llm.constants
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pykoi.llm.huggingface module
-----------------------------
-
-.. automodule:: pykoi.llm.huggingface
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pykoi.llm.model\_factory module
--------------------------------
-
-.. automodule:: pykoi.llm.model_factory
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pykoi.llm.openai module
------------------------
-
-.. automodule:: pykoi.llm.openai
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pykoi.llm.peft\_huggingface module
-----------------------------------
-
-.. automodule:: pykoi.llm.peft_huggingface
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pykoi.llm
-   :members:
-   :undoc-members:
-   :show-inheritance:
