+++ docs/pykoi.telemetry.rst
@@ -1,0 +1,29 @@
+pykoi.telemetry package
+=======================
+
+Submodules
+----------
+
+pykoi.telemetry.events module
+-----------------------------
+
+.. automodule:: pykoi.telemetry.events
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pykoi.telemetry.telemetry module
+--------------------------------
+
+.. automodule:: pykoi.telemetry.telemetry
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pykoi.telemetry
+   :members:
+   :undoc-members:
+   :show-inheritance:
