+++ docs/tests.chat.db.rst
@@ -1,0 +1,29 @@
+tests.chat.db package
+=====================
+
+Submodules
+----------
+
+tests.chat.db.test\_qa\_database module
+---------------------------------------
+
+.. automodule:: tests.chat.db.test_qa_database
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.chat.db.test\_ranking\_database module
+--------------------------------------------
+
+.. automodule:: tests.chat.db.test_ranking_database
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: tests.chat.db
+   :members:
+   :undoc-members:
+   :show-inheritance:
