+++ docs/tests.chat.llm.rst
@@ -1,0 +1,45 @@
+tests.chat.llm package
+======================
+
+Submodules
+----------
+
+tests.chat.llm.test\_abs\_llm module
+------------------------------------
+
+.. automodule:: tests.chat.llm.test_abs_llm
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.chat.llm.test\_constants module
+-------------------------------------
+
+.. automodule:: tests.chat.llm.test_constants
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.chat.llm.test\_huggingface module
+---------------------------------------
+
+.. automodule:: tests.chat.llm.test_huggingface
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.chat.llm.test\_openai module
+----------------------------------
+
+.. automodule:: tests.chat.llm.test_openai
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: tests.chat.llm
+   :members:
+   :undoc-members:
+   :show-inheritance:
