+++ docs/tests.chat.rst
@@ -1,0 +1,19 @@
+tests.chat package
+==================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   tests.chat.db
+   tests.chat.llm
+
+Module contents
+---------------
+
+.. automodule:: tests.chat
+   :members:
+   :undoc-members:
+   :show-inheritance:
