+++ docs/tests.component.rst
@@ -1,0 +1,21 @@
+tests.component package
+=======================
+
+Submodules
+----------
+
+tests.component.test\_base module
+---------------------------------
+
+.. automodule:: tests.component.test_base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: tests.component
+   :members:
+   :undoc-members:
+   :show-inheritance:
