+++ pykoi/__init__.py
@@ -1,10 +1,10 @@
 from pykoi.application import Application
 from pykoi.component.base import Chatbot, Dashboard, Dropdown
 from pykoi.component.chatbot_comparator import Compare
-from pykoi.db.qa_database import QuestionAnswerDatabase
-from pykoi.db.ranking_database import RankingDatabase
-from pykoi.llm.abs_llm import AbsLlm
-from pykoi.llm.model_factory import ModelFactory
+from pykoi.chat.db.qa_database import QuestionAnswerDatabase
+from pykoi.chat.db.ranking_database import RankingDatabase
+from pykoi.chat.llm.abs_llm import AbsLlm
+from pykoi.chat.llm.model_factory import ModelFactory
 from pykoi.rlhf.supervised_finetuning import SupervisedFinetuning
 from pykoi.rlhf.rw_finetuning import RewardFinetuning
 from pykoi.rlhf.rl_finetuning import RLFinetuning
