+++ pykoi/chat/db/abs_database.py
