+++ pykoi/chat/db/constants.py
