+++ pykoi/component/chatbot_comparator.py
@@ -5,11 +5,11 @@ import pandas as pd
 from typing import List
 
 from pykoi.component.base import Component
-from pykoi.db.comparator_database import (
+from pykoi.chat.db.comparator_database import (
     ComparatorDatabase,
     ComparatorQuestionDatabase,
 )
-from pykoi.llm.abs_llm import AbsLlm
+from pykoi.chat.llm.abs_llm import AbsLlm
 from pykoi.interactives.barchart import Barchart
