+++ pykoi/component/chatbot_database_factory.py
@@ -2,8 +2,8 @@
 from typing import Union
 
 from pykoi.component.constants import FeedbackType
-from pykoi.db.qa_database import QuestionAnswerDatabase
-from pykoi.db.ranking_database import RankingDatabase
+from pykoi.chat.db.qa_database import QuestionAnswerDatabase
+from pykoi.chat.db.ranking_database import RankingDatabase
 
 
 class ChatbotDatabaseFactory:
