+++ tests/chat/db/test_ranking_database.py
@@ -5,7 +5,7 @@ import os
 import sqlite3
 import unittest
 
-from pykoi.db.ranking_database import RankingDatabase
+from pykoi.chat.db.ranking_database import RankingDatabase
 
 # Define a temporary database file for testing
 TEST_DB_FILE = "test_ranking.db"
